test(plugin-flow-builder): move flows folder into helpers folder
